add missing may-block checks for sync-to-sync guest-to-guest calls
Previously, we weren't updating or checking the may-block status of a task
across sync-to-sync, guest-to-guest calls, meaning we were allowing blocking in
cases we shouldn't have.

This fixes that by adding a new `task_may_block` field to `VMComponentContext`,
plus code to update it every time we switch threads or do a sync-to-sync,
guest-to-guest call.  We use that field as the source of truth about whether a
blocking operation is permitted.

I've updated various tests to match, and Luke has an item on his to-do list to
add sad-path coverage for various cases to the upstream `component-model` test
suite.

+++ crates/wasmtime/src/runtime/component/concurrent.rs
@@ -58,7 +58,7 @@ use crate::component::{
 use crate::fiber::{self, StoreFiber, StoreFiberYield};
 use crate::store::{Store, StoreId, StoreInner, StoreOpaque, StoreToken};
 use crate::vm::component::{
-    CallContext, ComponentInstance, HandleTable, InstanceFlags, ResourceTables,
+    CallContext, ComponentInstance, HandleTable, InstanceFlags, ResourceTables, TaskMayBlock,
 };
 use crate::vm::{AlwaysMut, SendSyncPtr, VMFuncRef, VMMemoryDefinition, VMStore};
 use crate::{
@@ -710,12 +710,6 @@ pub(crate) enum WaitResult {
     Completed,
 }
 
-/// Raise a trap if the calling task is synchronous and trying to block prior to
-/// returning a value.
-pub(crate) fn check_blocking(store: &mut dyn VMStore) -> Result<()> {
-    store.concurrent_state_mut().check_blocking()
-}
-
 /// Poll the specified future until it completes on behalf of a guest->host call
 /// using a sync-lowered import.
 ///
@@ -841,7 +835,7 @@ fn handle_guest_call(store: &mut dyn VMStore, call: GuestCall) -> Result<()> {
                     call.thread,
                 );
 
-                let old_thread = state.guest_thread.replace(call.thread);
+                let old_thread = store.set_thread(Some(call.thread));
                 log::trace!(
                     "GuestCallKind::DeliverEvent: replaced {old_thread:?} with {:?} as current thread",
                     call.thread
@@ -869,6 +863,8 @@ fn handle_guest_call(store: &mut dyn VMStore, call: GuestCall) -> Result<()> {
 
                 store.maybe_pop_call_context(call.thread.task)?;
 
+                store.set_thread(old_thread);
+
                 next = instance.handle_callback_code(
                     store,
                     call.thread,
@@ -876,7 +872,6 @@ fn handle_guest_call(store: &mut dyn VMStore, call: GuestCall) -> Result<()> {
                     code,
                 )?;
 
-                store.concurrent_state_mut().guest_thread = old_thread;
                 log::trace!(
                     "GuestCallKind::DeliverEvent: restored {old_thread:?} as current thread"
                 );
@@ -1404,7 +1399,6 @@ impl StoreOpaque {
         StoreComponentInstanceId::new(self.id(), instance.instance)
             .get_mut(self)
             .instance_state(instance.index)
-            .unwrap()
             .concurrent_state()
     }
 
@@ -1414,8 +1408,60 @@ impl StoreOpaque {
         StoreComponentInstanceId::new(self.id(), instance.instance)
             .get_mut(self)
             .instance_state(instance.index)
-            .unwrap()
             .handle_table()
+    }
+
+    fn task_may_block(&mut self, instance: ComponentInstanceId) -> TaskMayBlock {
+        StoreComponentInstanceId::new(self.id(), instance)
+            .get_mut(self)
+            .task_may_block()
+    }
+
+    fn set_thread(&mut self, thread: Option<QualifiedThreadId>) -> Option<QualifiedThreadId> {
+        // Each time we switch threads, we conservatively set `task_may_block`
+        // to `false` for the component instance we're switching away from (if
+        // any), meaning it will be `false` for any new thread created for that
+        // instance unless explicitly set otherwise.
+        let state = self.concurrent_state_mut();
+        let old_thread = state.guest_thread.take();
+        if let Some(old_thread) = old_thread {
+            let instance = state.get_mut(old_thread.task).unwrap().instance.instance;
+            let mut task_may_block = self.task_may_block(instance);
+            unsafe { task_may_block.set(false) }
+        }
+
+        self.concurrent_state_mut().guest_thread = thread;
+
+        // If we're switching to a new thread, set its component instance's
+        // `task_may_block` according to where it left off.
+        if thread.is_some() {
+            self.set_task_may_block();
+        }
+
+        old_thread
+    }
+
+    /// Set the global variable representing whether the current task may block
+    /// prior to entering Wasm code.
+    fn set_task_may_block(&mut self) {
+        let state = self.concurrent_state_mut();
+        let guest_thread = state.guest_thread.unwrap();
+        let instance = state.get_mut(guest_thread.task).unwrap().instance.instance;
+        let mut task_may_block = self.task_may_block(instance);
+        let may_block = self.concurrent_state_mut().may_block(guest_thread.task);
+        unsafe { task_may_block.set(may_block) }
+    }
+
+    pub(crate) fn check_blocking(&mut self) -> Result<()> {
+        let state = self.concurrent_state_mut();
+        let task = state.guest_thread.unwrap().task;
+        let instance = state.get_mut(task).unwrap().instance.instance;
+        let task_may_block = self.task_may_block(instance);
+        if unsafe { task_may_block.get() } {
+            Ok(())
+        } else {
+            Err(Trap::CannotBlockSyncTask.into())
+        }
     }
 
     /// Record that we're about to enter a (sub-)component instance which does
@@ -1483,9 +1529,10 @@ impl StoreOpaque {
 
         let fiber = fiber::resolve_or_release(self, fiber).await?;
 
+        self.set_thread(old_thread);
+
         let state = self.concurrent_state_mut();
 
-        state.guest_thread = old_thread;
         if let Some(ref ot) = old_thread {
             state.get_mut(ot.thread)?.state = GuestThreadState::Running;
         }
@@ -1579,7 +1626,7 @@ impl StoreOpaque {
         self.with_blocking(|_, cx| cx.suspend(StoreFiberYield::ReleaseStore))?;
 
         if let Some(task) = task {
-            self.concurrent_state_mut().guest_thread = old_guest_thread;
+            self.set_thread(old_guest_thread);
             self.maybe_push_call_context(task)?;
         }
 
@@ -1948,10 +1995,7 @@ impl Instance {
                     store,
                     callee_instance.index,
                 )?;
-                let old_thread = store
-                    .concurrent_state_mut()
-                    .guest_thread
-                    .replace(guest_thread);
+                let old_thread = store.set_thread(Some(guest_thread));
                 log::trace!(
                     "stackless call: replaced {old_thread:?} with {guest_thread:?} as current thread"
                 );
@@ -1972,8 +2016,8 @@ impl Instance {
 
                 store.maybe_pop_call_context(guest_thread.task)?;
 
+                store.set_thread(old_thread);
                 let state = store.concurrent_state_mut();
-                state.guest_thread = old_thread;
                 old_thread
                     .map(|t| state.get_mut(t.thread).unwrap().state = GuestThreadState::Running);
                 log::trace!("stackless call: restored {old_thread:?} as current thread");
@@ -1993,10 +2037,7 @@ impl Instance {
                     store,
                     callee_instance.index,
                 )?;
-                let old_thread = store
-                    .concurrent_state_mut()
-                    .guest_thread
-                    .replace(guest_thread);
+                let old_thread = store.set_thread(Some(guest_thread));
                 log::trace!(
                     "sync/async-stackful call: replaced {old_thread:?} with {guest_thread:?} as current thread",
                 );
@@ -2104,6 +2145,8 @@ impl Instance {
                     )?;
                 }
 
+                store.set_thread(old_thread);
+
                 store.maybe_pop_call_context(guest_thread.task)?;
 
                 let state = store.concurrent_state_mut();
@@ -2166,7 +2209,7 @@ impl Instance {
             // A task may only call an async-typed function via a sync lower if
             // it was created by a call to an async export.  Otherwise, we'll
             // trap.
-            store.0.concurrent_state_mut().check_blocking()?;
+            store.0.check_blocking()?;
         }
 
         enum ResultInfo {
@@ -2327,10 +2370,10 @@ impl Instance {
 
         // Make the new thread the current one so that `Self::start_call` knows
         // which one to start.
-        state.guest_thread = Some(QualifiedThreadId {
+        store.0.set_thread(Some(QualifiedThreadId {
             task: guest_task,
             thread: guest_thread,
-        });
+        }));
         log::trace!(
             "pushed {guest_task:?}:{guest_thread:?} as current thread; old thread was {old_thread:?}"
         );
@@ -2544,32 +2587,29 @@ impl Instance {
             }
         };
 
-        let state = store.0.concurrent_state_mut();
+        guest_waitable.join(store.0.concurrent_state_mut(), old_set)?;
 
-        guest_waitable.join(state, old_set)?;
+        // Reset the current thread to point to the caller as it resumes control.
+        store.0.set_thread(Some(caller));
+        store.0.concurrent_state_mut().get_mut(caller.thread)?.state = GuestThreadState::Running;
+        log::trace!("popped current thread {guest_thread:?}; new thread is {caller:?}");
 
         if let Some(storage) = storage {
             // The caller used a sync-lowered import to call an async-lifted
             // export, in which case the result, if any, has been stashed in
             // `GuestTask::sync_result`.
+            let state = store.0.concurrent_state_mut();
             let task = state.get_mut(guest_thread.task)?;
             if let Some(result) = task.sync_result.take() {
                 if let Some(result) = result {
                     storage[0] = MaybeUninit::new(result);
                 }
 
-                if task.exited {
-                    if task.ready_to_delete() {
-                        Waitable::Guest(guest_thread.task).delete_from(state)?;
-                    }
+                if task.exited && task.ready_to_delete() {
+                    Waitable::Guest(guest_thread.task).delete_from(state)?;
                 }
             }
         }
-
-        // Reset the current thread to point to the caller as it resumes control.
-        state.guest_thread = Some(caller);
-        state.get_mut(caller.thread)?.state = GuestThreadState::Running;
-        log::trace!("popped current thread {guest_thread:?}; new thread is {caller:?}");
 
         Ok(status.pack(waitable))
     }
@@ -3015,7 +3055,7 @@ impl Instance {
             // The caller may only call `waitable-set.wait` from an async task
             // (i.e. a task created via a call to an async export).
             // Otherwise, we'll trap.
-            store.concurrent_state_mut().check_blocking()?;
+            store.check_blocking()?;
         }
 
         let &CanonicalOptions {
@@ -3122,10 +3162,7 @@ impl Instance {
         let token = StoreToken::new(store.as_context_mut());
         let start_func = Box::new(
             move |store: &mut dyn VMStore, guest_thread: QualifiedThreadId| -> Result<()> {
-                let old_thread = store
-                    .concurrent_state_mut()
-                    .guest_thread
-                    .replace(guest_thread);
+                let old_thread = store.set_thread(Some(guest_thread));
                 log::trace!(
                     "thread start: replaced {old_thread:?} with {guest_thread:?} as current thread"
                 );
@@ -3147,7 +3184,8 @@ impl Instance {
                 if task.threads.is_empty() && !task.returned_or_cancelled() {
                     bail!(Trap::NoAsyncResult);
                 }
-                state.guest_thread = old_thread;
+                store.0.set_thread(old_thread);
+                let state = store.0.concurrent_state_mut();
                 old_thread
                     .map(|t| state.get_mut(t.thread).unwrap().state = GuestThreadState::Running);
                 if state.get_mut(guest_thread.task)?.ready_to_delete() {
@@ -3256,7 +3294,7 @@ impl Instance {
                 // The caller may only call `thread.suspend` from an async task
                 // (i.e. a task created via a call to an async export).
                 // Otherwise, we'll trap.
-                state.check_blocking()?;
+                store.check_blocking()?;
             }
         }
 
@@ -3397,7 +3435,7 @@ impl Instance {
             // The caller may only sync call `subtask.cancel` from an async task
             // (i.e. a task created via a call to an async export).  Otherwise,
             // we'll trap.
-            store.concurrent_state_mut().check_blocking()?;
+            store.check_blocking()?;
         }
 
         let (rep, is_host) = store
@@ -4937,11 +4975,6 @@ impl ConcurrentState {
         } else {
             false
         }
-    }
-
-    fn check_blocking(&mut self) -> Result<()> {
-        let task = self.guest_thread.unwrap().task;
-        self.check_blocking_for(task)
     }
 
     fn check_blocking_for(&mut self, task: TableId<GuestTask>) -> Result<()> {
